requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
lightgbm
shap
streamlit
joblib
torch
torchvision
pillow